pyproject.toml
[project]
name = "finance-reranker"
version = "0.1.0"
description = "Fine-tuning a cross-encoder reranker for financial document retrieval."
requires-python = ">=3.11"
dependencies = [
  "torch",
  "transformers",
  "datasets",
  "peft",
  "accelerate",
  "bitsandbytes",
  "pyyaml",
  "pydantic",
  "pandas",
  "numpy",
  "scikit-learn",
  "tqdm",
  "jupyter",
  "unsloth>=2026.3.17",
  "trl>=0.24.0",
  "sentence-transformers>=5.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
